% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-data.R
\name{as_shared_tsibble}
\alias{as_shared_tsibble}
\title{Coerce to a shared tsibble from tsibble}
\usage{
as_shared_tsibble(x, spec)
}
\arguments{
\item{x}{A tsibble.}

\item{spec}{A formula to specify tsibble key structures. By default, crossing
structures (i.e \code{key1 * key2}) are assumed for the key. The required
specification for nesting is \code{parent / child}.}
}
\value{
\code{SharedTsibbleData} subclassing of \code{SharedData} from {crosstalk}
}
\description{
Coerce to a shared tsibble from tsibble
}
\details{
The object returned from \code{as_shared_tsibble()} can be piped into all {plotly}
functions to be visualised.
}
\examples{
library(tsibble)
as_shared_tsibble(tourism, spec = (State / Region) * Purpose)
}
