% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{plotly_key_tree}
\alias{plotly_key_tree}
\title{Plot nesting structures in shared tsibbles using plotly}
\usage{
plotly_key_tree(data, height = NULL, width = NULL, ...)
}
\arguments{
\item{data}{A shared tsibble.}

\item{height}{height}

\item{width}{width}

\item{...}{arguments supplied to \code{\link[plotly:subplot]{subplot()}}}
}
\description{
Plot nesting structures in shared tsibbles using plotly
}
