% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sunspots2019}
\alias{sunspots2019}
\title{Yearly mean total sunspot number (1700 - 2019)}
\format{
An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 320 rows and 2 columns.
}
\usage{
sunspots2019
}
\description{
Yearly mean total sunspot number (1700 - 2019)
}
\examples{
data(sunspots2019)
}
\references{
\href{http://www.sidc.be/silso/datafiles}{WDC-SILSO, Royal Observatory of Belgium, Brussels}
}
\keyword{datasets}
