% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tourism_monthly}
\alias{tourism_monthly}
\title{Monthly Australian domestic overnight trips}
\format{
A tsibble with 80,696 rows and 5 variables:
\itemize{
\item \strong{Month}: Year month (index)
\item \strong{State}: States and territories of Australia
\item \strong{Region}: The tourism regions are formed through the aggregation of
Statistical Local Areas (SLAs) which are defined by the various State and
Territory tourism authorities according to their research and marketing
needs
\item \strong{Purpose}: Stopover purpose of visit:
\itemize{
\item "Holiday"
\item "Visiting friends and relatives"
\item "Business"
\item "Other reason"
}
\item \strong{Trips}: Overnight trips in thousands
}
}
\usage{
tourism_monthly
}
\description{
A dataset containing the monthly overnight trips from 1998 Jan to 2019 Dec
across Australia.
}
\examples{
data(tourism_monthly)
}
\references{
\href{https://www.tra.gov.au}{Tourism Research Australia}
}
\keyword{datasets}
