% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibbletalk-package.R
\docType{package}
\name{tsibbletalk-package}
\alias{tsibbletalk}
\alias{tsibbletalk-package}
\title{tsibbletalk: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Earo Wang \email{earo.wang@gmail.com} (\href{https://orcid.org/0000-0001-6448-5260}{ORCID})

}
\keyword{internal}
