% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module.R
\name{tsibble-dice}
\alias{tsibble-dice}
\alias{tsibbleDiceUI}
\alias{tsibbleDiceServer}
\title{A shiny module to easily slice and dice tsibble index for visualising periodicity}
\usage{
tsibbleDiceUI(id)

tsibbleDiceServer(id, plot, period)
}
\arguments{
\item{id}{A unique shiny id.}

\item{plot}{A \code{ggplot} or \code{plotly} object.}

\item{period}{A string passed to \code{\link[lubridate:period]{lubridate::period()}} to specify the minimul
seasonal period, for example \code{"1 day"}.}
}
\description{
A pair of UI and server functions: \code{tsibbleDiceUI()} and \code{tsibbleDiceServer()}.
}
