% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibbletalk-package.R
\docType{package}
\name{tsibbletalk-package}
\alias{tsibbletalk}
\alias{tsibbletalk-package}
\title{tsibbletalk: Interactive Graphics for Tsibble Objects}
\description{
A shared tsibble data easily communicates between
    htmlwidgets on both client and server sides, powered by 'crosstalk'. A
    shiny module is provided to visually explore periodic/aperiodic
    temporal patterns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/earowang/tsibbletalk}
  \item Report bugs at \url{https://github.com/earowang/tsibbletalk/issues}
}

}
\author{
\strong{Maintainer}: Earo Wang \email{earo.wang@gmail.com} (\href{https://orcid.org/0000-0001-6448-5260}{ORCID})

Authors:
\itemize{
  \item Di Cook (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})
}

}
\keyword{internal}
